requirements.txt
alipay-sdk-python==3.6.72
aliyun-python-sdk-cms==7.0.32
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-ecs==4.24.20
APScheduler==3.9.1
asgiref==3.3.4
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-apscheduler==0.6.2
django-braces==1.15.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==21.1
django-mdeditor==0.1.20
django-oauth-toolkit==1.5.0
django-redis==5.2.0
django-simpleui==2022.4.9
djangorestframework==3.13.1
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.1.0
drf-extensions==0.7.1
idna==2.10
image==1.5.33
jmespath==0.10.0
jwcrypto==1.0
loguru==0.5.3
lxml==4.6.3
mysqlclient==2.1.0
oauthlib==3.1.1
optionaldict==0.1.2
packaging==21.3
Pillow==9.1.1
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.12.0
PyJWT==2.1.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2022.1
qiniu==7.4.1
qrcode==7.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.1
tencentcloud-sdk-python==3.0.652
tzlocal==2.1
urllib3==1.26.5
wrapt==1.13.3
